Remove keylistener in favor of notelistener
- note listener takes into account press and release, so it can handle chords
- the notes played dispatcher dispatches a set of notes each time a note is pressed/released
